[#47] refactor: SideDishException 클래스 추가
