+++ backend/src/main/java/com/terria/sidedish/error/ErrorCode.java
@@ -5,7 +5,8 @@ import org.springframework.http.HttpStatus;
 public enum ErrorCode {
 
     EXHIBITION_VALIDATION_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "기획전 아이디가 유효하지 않습니다."),
-    NO_SUCH_MEMBER_ERROR(HttpStatus.NOT_FOUND, "해당 사용자가 존재하지 않습니다.");
+    NO_SUCH_MEMBER_ERROR(HttpStatus.NOT_FOUND, "해당 사용자가 존재하지 않습니다."),
+    NO_SUCH_SIDE_DISH_ERROR(HttpStatus.NOT_FOUND, "해당 사용자가 존재하지 않습니다.");
 
     final HttpStatus status;
     final String message;
