+++ backend/src/main/java/com/terria/sidedish/error/HandlerAdvice.java
@@ -1,5 +1,8 @@
 package com.terria.sidedish.error;
 
+import com.terria.sidedish.error.exception.ExhibitionException;
+import com.terria.sidedish.error.exception.OAuthException;
+import com.terria.sidedish.error.exception.SideDishException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -23,8 +26,20 @@ public class HandlerAdvice {
         return new ResponseEntity<>(new ErrorResponse<>(messages), HttpStatus.BAD_REQUEST);
     }
 
+    @ExceptionHandler(OAuthException.class)
+    private ResponseEntity<ErrorResponse<String>> handleOAuthException(OAuthException e) {
+        ErrorCode errorCode = e.getErrorCode();
+        return new ResponseEntity<>(new ErrorResponse<>(errorCode.getMessage()), errorCode.getStatus());
+    }
+
     @ExceptionHandler(ExhibitionException.class)
-    private ResponseEntity<ErrorResponse<String>> handleCardRuntimeException(ExhibitionException e) {
+    private ResponseEntity<ErrorResponse<String>> handleExhibitionException(ExhibitionException e) {
+        ErrorCode errorCode = e.getErrorCode();
+        return new ResponseEntity<>(new ErrorResponse<>(errorCode.getMessage()), errorCode.getStatus());
+    }
+
+    @ExceptionHandler(SideDishException.class)
+    private ResponseEntity<ErrorResponse<String>> handleSideDishException(SideDishException e) {
         ErrorCode errorCode = e.getErrorCode();
         return new ResponseEntity<>(new ErrorResponse<>(errorCode.getMessage()), errorCode.getStatus());
     }
