+++ backend/src/main/java/com/terria/sidedish/service/GitHubAuthService.java
@@ -5,8 +5,8 @@ import com.terria.sidedish.dto.auth.GitHubAccessToken;
 import com.terria.sidedish.dto.auth.GitHubUser;
 import com.terria.sidedish.dto.auth.Provider;
 import com.terria.sidedish.dto.response.MemberResponse;
-import com.terria.sidedish.error.GitHubOAuthException;
-import com.terria.sidedish.error.OAuthException;
+import com.terria.sidedish.error.exception.GitHubOAuthException;
+import com.terria.sidedish.error.exception.OAuthException;
 import com.terria.sidedish.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
