+++ backend/src/main/java/com/terria/sidedish/service/SideDishService.java
@@ -6,6 +6,8 @@ import com.terria.sidedish.domain.entity.reference.DiscountEvent;
 import com.terria.sidedish.dto.response.sidedish.SideDishCardResponse;
 import com.terria.sidedish.dto.response.sidedish.SideDishCardResponses;
 import com.terria.sidedish.dto.response.sidedish.SideDishDetailResponse;
+import com.terria.sidedish.error.ErrorCode;
+import com.terria.sidedish.error.exception.SideDishException;
 import com.terria.sidedish.repository.DiscountEventRepository;
 import com.terria.sidedish.repository.SideDishRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +17,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-
 
 @Service
 @RequiredArgsConstructor
@@ -39,7 +40,7 @@ public class SideDishService {
 
     public SideDishDetailResponse getById(long id) {
         SideDish sideDish = sideDishRepository.findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new SideDishException(ErrorCode.NO_SUCH_SIDE_DISH_ERROR));
 
         return SideDishDetailResponse.from(sideDish, getDiscountEvent(sideDish));
     }
